renamed enumspecindex to wildcardindex
