+++ tsconfig.json
@@ -42,7 +42,7 @@
         "src/ranking/ranking.ts",
         "src/config.ts",
         "src/enumerator.ts",
-        "src/enumspecindex.ts",
+        "src/wildcardindex.ts",
         "src/wildcard.ts",
         "src/generate.ts",
         "src/model.ts",
@@ -66,7 +66,7 @@
         "test/ranking/rule.ts",
         "test/enumerator.test.ts",
         "test/wildcard.test.ts",
-        "test/enumspecindex.test.ts",
+        "test/wildcardindex.test.ts",
         "test/generate.test.ts",
         "test/model.test.ts",
         "test/nest.test.ts",
